Split native convertion functions into separate files
