+++ packages/realm_dart/lib/src/native/mutable_subscription_set_handle.dart
@@ -6,9 +6,9 @@ import 'dart:ffi';
 import 'package:ffi/ffi.dart';
 
 import '../realm_dart.dart'; // TODO: Remove this import
+import 'convert_native.dart';
 import 'error_handling.dart';
 import 'realm_bindings.dart';
-import 'realm_core.dart'; // TODO: Remove this import
 import 'realm_handle.dart';
 import 'realm_library.dart';
 import 'results_handle.dart';
