+++ packages/realm_dart/lib/src/native/realm_handle.dart
@@ -12,12 +12,12 @@ import '../logging.dart';
 import '../realm_class.dart';
 import '../realm_object.dart';
 import 'config_handle.dart';
+import 'convert_native.dart';
 import 'error_handling.dart';
 import 'handle_base.dart';
 import 'object_handle.dart';
 import 'query_handle.dart';
 import 'realm_bindings.dart';
-import 'realm_core.dart';
 import 'realm_library.dart';
 import 'results_handle.dart';
 import 'rooted_handle.dart';
