+++ packages/realm_dart/lib/src/native/results_handle.dart
@@ -6,6 +6,7 @@ import 'dart:ffi';
 import 'package:ffi/ffi.dart';
 
 import '../realm_dart.dart';
+import 'convert_native.dart';
 import 'error_handling.dart';
 import 'notification_token_handle.dart';
 import 'object_handle.dart';
