+++ packages/realm_dart/lib/src/native/session_handle.dart
@@ -9,13 +9,14 @@ import 'package:ffi/ffi.dart';
 import '../realm_dart.dart';
 import '../scheduler.dart';
 import '../session.dart'; // TODO: Remove this import
+import 'convert_native.dart';
 import 'handle_base.dart';
 import 'realm_bindings.dart';
-import 'realm_core.dart'; // TODO: Remove this import
 import 'realm_handle.dart';
 import 'realm_library.dart';
 import 'rooted_handle.dart';
 import 'user_handle.dart';
+
 
 class SessionHandle extends RootedHandleBase<realm_sync_session_t> {
   @override
