+++ packages/realm_dart/lib/src/native/set_handle.dart
@@ -6,6 +6,7 @@ import 'dart:ffi';
 import 'package:ffi/ffi.dart';
 
 import '../realm_dart.dart'; // TODO: remove this import
+import 'convert_native.dart';
 import 'error_handling.dart';
 import 'notification_token_handle.dart';
 import 'query_handle.dart';
